refactor(fast-usdc): ContractMeta, Permit
Factor out metainfo so that generic .start.js can get it without
getting the whole contract.
